finish refactoring home section
